# Include all language tags, not just two letter tags