--- a/test/integration/rules/valid-lang/valid-lang.html
+++ b/test/integration/rules/valid-lang/valid-lang.html
@@ -4,6 +4,7 @@
 <p xml:lang="en-US" lang="en-US" id="pass3">Mix</p>
 <p xml:lang="de" lang="de" id="pass4">Deutsch</p>
 <p lang="af" id="pass5">Afrikaans</p>
+<p lang="oma" id="pass6">Omaha (3 letter)</p>
 
 <p lang="gibberish-US" id="fail1">Not English</p>
 <p xml:lang="gibberish-US" id="fail2">Not English</p>