--- a/test/integration/rules/valid-lang/valid-lang.json
+++ b/test/integration/rules/valid-lang/valid-lang.json
@@ -2,5 +2,5 @@
 	"description": "valid-lang test",
 	"rule": "valid-lang",
 	"violations": [["#fail1"], ["#fail2"], ["#fail3"]],
-	"passes": [["#pass1"], ["#pass2"], ["#pass3"], ["#pass4"], ["#pass5"]]
+	"passes": [["#pass1"], ["#pass2"], ["#pass3"], ["#pass4"], ["#pass5"], ["#pass6"]]
 }