While testing several sites, I came across one about the Omaha language. The site used the 'oma' language tag to mark up Omaha language on the site, which appears to conform to BCP47 and is listed in the IANA registry. However, aXe-core was flagging the language as an error because it is only pulling in two character language codes from the registry. I may be missing something, but I don't see anything in the HTML or WCAG standards that limit usage to 2 character codes.