@dylanb 2016-11-22
@mfairchild365 where did you find the language code in the registry? Can you post a link here?