@mfairchild365 2016-11-22
Sure, http://www.iana.org/assignments/language-subtag-registry/language-subtag-registry and this is the relevant record:

```
Type: language
Subtag: oma
Description: Omaha-Ponca
Added: 2009-07-29
```