@mfairchild365 2016-11-22
@dylanb I added an integration test to cover this case too.

I should mention that I seriously doubt any AT would know how to handle the oma language, but that doesn't change that it is still a valid language code.