@dylanb 2016-11-22
@mfairchild365 you will need to run the language generation task (it does not run every time). This will update the `valid-lang.json` file and then you need to submit that file as part of your PR. I would like to see how much bigger that makes our final code before deciding on whether to approve this change.