@dylanb 2016-11-22
Wow, this is increasing the minified file size by 62% from ~163K to 264K. The API already has a way to add/change languages. I would prefer to add a capability to the UI of the extension to allow specific languages to be added rather than bloat the library for languages that are not supported by any ATs.

@WilcoFiers @marcysutton WDYT?