@marcysutton 2016-11-22
I would also prefer to not add much to the filesize if we can enable language support through the API and an affordance in the extension UI–especially if the overhead doesn't gain much for AT. The bulk of the content added is in `valid-lang.json`–how do we decide which languages to include? The more common ones?