@mfairchild365 2016-11-22
While I understand where you are coming from, I disagree. In my opinion (and I realize that I may not share the same opinions as the project maintainers), I see the file size of the library as not that important when compared to the items listed in the project manifesto, specifically a zero false positive test rate.

If the majority of users are using this library as a browser extension, they only need to download the library once (or whenever it is updated), so why not just include every valid language? The scenario of automated testing with the library (integration or a crawling method), might benefit more from a smaller library, but again, I don't think file size should get in the way of the core objectives of the library (but a small file size should a goal).