@boris1558 2016-11-22
I wonder if loading a common language set and then doing a query to http://www.iana.org/assignments/language-subtag-registry/language-subtag-registry would accomplish both?