@mfairchild365 2016-11-23
@boris1558 performing a query to the registry would add an external dependency, which is an extra point of failure and might not even be possible in closed environments. However, it is a possibility.

@dylanb I had a think about this some more and ended up doing some more research. What do you think about testing the structure of the language tag instead of testing the value against a predefined set? This could be accomplished through something like https://www.npmjs.com/package/bcp-47

Advantages to testing the structure:

- Much smaller file size impact
- Will not fail new languages that have been added to the registry since we last updated our code
- Can validate the entire language tag (language, scirpt, region, variants, etc). Although I'm not sure how much of an impact this would have on AT. Would AT toss out the entire language tag if the script part was in an invalid format?

Disadvantages to testing the structure

- Possibility of false negatives by not detecting typos (typed 'em' instead of 'en') 

I also did some research regarding how other tools handle this.

- The i18n JavaScript API appears to test the structure: http://ecma-international.org/ecma-402/1.0/#sec-6.2.2
- HTML_Codesniffer tests the structure: https://github.com/squizlabs/HTML_CodeSniffer/blob/9d6832f4f332f21c2b63899f4e9a804a381237e2/Standards/WCAG2AAA/Sniffs/Principle3/Guideline3_1/3_1_1.js#L70
- It looks like the w3c validator checks both structure and value: https://github.com/validator/validator/blob/fe88da45fd505a4add9a61d1e15c3a57cd8b30fe/src/nu/validator/datatype/Language.java#L109