@dylanb 2016-11-23
@mfairchild365 I agree - external call is not the way we want to go for the reasons you mention PLUS it violates our manifesto in that it would introduce lag and failure modes we do not want to introduce.

@marcysutton given that our library is NOT really a run-time library for most use cases, I think I am leaning towards suggesting that we merge this change and if we get compelling evidence that it is causing problems, consider actions based on the details of those use cases. Before we do that, I would like to test to see what the actual performance impact is when the large library is included into our extension(s). Could you do that test (a before and after test)?