@marcysutton 2016-11-23
I do have benchmarking on my plate as part of our performance improvements. But we have to weigh priority against the ES6 architecture spike I was about to start. Maybe I should try to finish the perf improvements before moving on to architecture?