@dylanb 2016-11-24
yes, that sounds like the right priorities