@marcysutton 2016-12-19
I ran 2 axe audits with our performance timer on different sites with and without this PR, and it was a very small increase. The filesize is honestly a bigger issue than what I found. If you have another page you want me to test, let me know. I'm still working on automating this process so we can have it run against PRs automatically.

## Mars Commuter page 
https://dequeuniversity.com/demo/mars/

**Before: all 4 frames**
Measure axe took 788.2050000000008ms

**After: all 4 frames**
Measure axe took 802.3150000000005ms

## MDN Promises page
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Promise

**Before: 2 frames**
Measure axe took 963.2800000000007ms

**After: 2 frames**
Measure axe took 996.3150000000005ms