@marcysutton 2016-12-20
@dylanb wanted more data on the effect on injecting a bigger file into axe-chrome, so I ran some performance measurements on that code:

## Mars Commuter page
https://dequeuniversity.com/demo/mars/

```
PerformanceMeasure
duration: 155.60499999999956
entryType: "measure"
name: "measure-startup"