@mfairchild365 2016-12-20
Interesting. I'm guessing the difference in start time is due to the larger file being loaded. I'm unsure of what exactly 'startTime' refers to. Is that the time is takes to start the tests after the 'Analyze' button is pressed?