@marcysutton 2016-12-20
I marked when a call is made to initialize frames in axe-chrome and again in that initializer's callback function. Then I measured those two marks. The measure has a start time and a duration, and the former definitely did look impacted by the file size.

From the User Timing API: `The startTime attribute will return a DOMHighResTimeStamp with the measure's start mark ` - https://www.w3.org/TR/user-timing/