@marcysutton 2016-12-22
We decided that the duration difference was minimal enough that it was worth the change, despite the filesize increase. And startTime is irrelevant. 

Thanks so much for your contribution!